api-3/master: Add include to resolve missing types defined in reflection string_wrapper: Uses C API but include was missing api_3/tes_array: Test code replacing non-existing VMArray (+ include fixes) api_3/tes_atomic: Fix typo and missing includes api_3/tes_map: master include was missing api_3/tes_object: add missing includes TES_Form is now inside namespace RE (CommonLibNG-SE) collections/collection.h: Move all types into separate include files   Fix old way to call this template method collections/collections_types.h: All types moved to here Rename several variables to fix invalid shadow template handling collections/item.h: Fix 32bit assumptions reflection/reflaction.h: Minor fix and one call commented (ToDo) Fix use of typenames in a few places Fix/add namespace of some other types for CommonLibNG-SE util/util.h: Comment boost namespace definition for the time being
