+++ src/JContainers/src/api_3/master.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <reflection/reflection.h>
+
 namespace tes_api_3 {
 
     template<class T>
