+++ src/JContainers/src/api_3/tes_array.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "collections/functions.h"
+#include "reflection/tes_binding.h"
+#include "master.h"
 
 namespace tes_api_3 {
 
@@ -58,8 +60,19 @@ namespace tes_api_3 {
         }
         REGISTERF2(objectWithSize, "size", "Creates a new array of given size, filled with empty (None) items");
 
+        template <class T>
+        class VMArrayCompat
+        {
+        public:
+            RE::BSScript::Array array;
+
+            std::uint32_t Length() const;
+            bool Get(T& out, std::uint32_t i);
+            bool Set(std::uint32_t i, const T& value);
+        };
+
         template<class TesType, class JCType = TesType>
-        static object_base* fromArray(tes_context& ctx, VMArray<TesType> arr)
+        static object_base* fromArray(tes_context& ctx, VMArrayCompat<TesType> arr)
         {
             JC_LOG_API ("...");
