+++ src/JContainers/src/collections/functions.h
@@ -96,7 +96,7 @@ namespace collections {
     struct map_key_checker {
         static bool check(const std::string& s)  { return !s.empty(); }
         static bool check(const char *s)  { return s != nullptr && *s; }
-        static bool check(TESForm *f)  { return f != nullptr; }
+        static bool check(RE::TESForm *f)  { return f != nullptr; }
         static bool check(FormId f)  { return f != FormId::Zero; }
         static bool check(const form_ref& f)  { return f.is_not_expired(); }
         static bool check(const form_ref_lightweight& f)  { return f.is_not_expired(); }
@@ -110,14 +110,14 @@ namespace collections {
         ///typedef typename T::key_type key_type;
 
         template<class Op, class R,/* class RAlter, */class key_type>
-        static R doReadOpR(T * obj, const key_type& key, R default, Op& operation) {
+        static R doReadOpR(T * obj, const key_type& key, R rDefault, Op& operation) {
             if (obj && key_checker::check(key)) {
                 object_lock g(obj);
                 item *itm = obj->u_get(key);
-                return itm ? operation(*itm) : default;
+                return itm ? operation(*itm) : rDefault;
             }
             else {
-                return default;
+                return rDefault;
             }
         }
 
@@ -161,8 +161,8 @@ namespace collections {
         }
 
         struct equal_to {
-            template<class T, class D>
-            inline bool operator()(T& t, D& d) const {
+            template<class cT, class cD>
+            inline bool operator()(cT& t, cD& d) const {
                 return t == d;
             }
         };
